backend/requirements.txt
asgiref==3.8.1
Django==5.2.7
djangorestframework==3.15.2
django-cors-headers==4.6.0
dj-database-url==3.0.1
django-extensions==4.1
django-filter==25.2
psycopg2-binary==2.9.10
pycparser==2.23
python-dotenv==1.0.1
requests==2.32.3
ruff==0.13.3
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.2
urllib3==1.26.16
Werkzeug==3.1.3
gunicorn==23.0.0
packaging==24.1
